POC: move generate_data() back from the SDG library
Now that generate_data() has been adapted to use the new SDG,
move it back into the CLI so that the CLI is only using the flexible
SDG API rather than continuing to have an all-in-one generate_data()
API with tight coupling to the CLI.

As an example of the tight coupling, CLI is calling generate_data()
and expecting it to produce train_* and test_* files. It would
seem more library like for the CLI to read the taxonomy and
convert it into the test_* file format, and then call
sdg.generate() and the generated samples into the train_* file
format.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>