generate: Allow pipeline selection
Some major enhancements have gone into the `instructlab/sdg` library
this week, including a more extensive pipeline that produces higher
quality results. This new pipeline requires a larger teacher model
than what we default to in the CLI, so it is optional and not used by
default.

This option provides the ability to turn it on by specifying
`--pipeline full`. The default value is `simple`, which maps to a
simple pipeline aimed at producing similar quality results as before,
though it is a new and different implementation. It's intended to be
good enough for basic testing and demos on a laptop or equivalent
small environment.

Signed-off-by: Russell Bryant <rbryant@redhat.com>