Introduce check_api_base
Adding "else:" to avoid indentation change and difficulties of review.

ensure_server() became technological debt.
Historically, it has aggregated a mix of functionally
different code, and its name does not accurately describe the content.
Also usage of this function is confusing.

The plan is to split the function and reorder the code accordingly
guidelines:
- clean backbend agnostic code
- clean llama-cpp related code
- clean vllm related code
- no code duplication
- modality

In order to refactoring, offshoot check_api_base() from ensure_server().

Signed-off-by: Costa Shulyupin <costa.shul@redhat.com>