Fix: No str.split or f-string for subprocess args
Never use `str.split` to split arguments to subprocess or f-strings to
create arguments for subprocess. This breaks as soon as there a
non-ASCII characters in any path, e.g. file names with a white space.

You either need `shlex.split` and `shlex.quote` (which does not work on
non-Unix shells) or use list of string all the way.

Also fix a typing bug with port. It's now an `int` all the way.

Signed-off-by: Christian Heimes <cheimes@redhat.com>