Add e2e CI variant that runs on AWS
This commit introduces a new approach for running the e2e CI job.
Instead of using a runner provided by github, use a similarly powered
runner on AWS that is much cheaper.

The real benefit of this approach is the ability to run on any other
instance type, giving us access to GPU instances that are more
powerful than what was built-in to GitHub. More powerful e2e variants
will come in future commits.

This job can only be manually triggered for now. At some point we may
create a configuration that can replace the one that runs
automatically on every commit, but I'd like to exercise this for a
while to make sure it's stable. It's also not quite as simple as just
running this automatically directly on PRs, because have to be careful
to protect against what's possible with a malicious PR. In any case,
that is future work.

Similar to how the original `e2e` job worked at first, this one can be
manually triggered to be run against any PR. Comments will be posted
to a PR so you can follow its progress, as well as be notified of its
results.

Signed-off-by: Russell Bryant <rbryant@redhat.com>
Co-authored-by: Brent Salisbury <bsalisbu@redhat.com>