ci: add option to run e2e test with the vLLM backend
The e2e script has now a new '-v' option (do not think verbose) that
will activate the vLLM backend when running the server.
We are sort of approaching the limit of what 'getopts' can do in terms
of choosing meaninful options. We could in the future transition to
'getopt' to allow for long flag to be added.

Part of  https://github.com/instructlab/instructlab/issues/1456

Signed-off-by: Sébastien Han <seb@redhat.com>
Co-authored-by: Russell Bryant <rbryant@redhat.com>