add backend flags to `ilab model evaluate` remove config editing
this new flag is defaulted from the _serve part of the config.yaml. If the user wants to, they can override the config options by using the new flag.

Also removed two edits to the config obj. Rather than editing the config, warn the user if the options in the config may be incompatible

Signed-off-by: Charlie Doern <cdoern@redhat.com>