generate: Add `--sdg-scale-factor`
The new implementation of data generation from the `instructlab/sdg`
library behaves differently than before. Instead of having a tunable
for the maximum number of generated examples, it has a tunable for the
number of examples to generate based on each seed example, for some
definition of "seed example". The number of seed examples expodes to
be much more than the provided q&a pairs when working with a knowledge
document.

The new option, `--sdg-scale-factor` tries to capture this new
behavior and make it clear that it's not a specific number of examples
anymore, but more of a multiplication factor.

We could come back and add a new option that behaves like the old one,
like `--max-generated-examples`, but even that is slightly different
than what was here before. Instead of implementing that now, hold off
until a good enough use case justifies the need for it, because it
would be harmful in some cases. For example, when working with a
knowledge document, setting a hard cap on the number of generated
examples will most likely halt the process before the full knowledge
document (or documents) have been processed.

The configuration file has also been updated to do the following:

- Optionally allow `sdg_scale_factor` so that previous config files
  still work. Many options have been added recently without doing
  this, breaking old configs.

- Allow num_instructions in the config file. This allows old config
  files to work without error, but a deprecation warning will be
  emitted.

- Ensure that the deprecated field is not included when generating
  a sample configuration file.

I believe this is the first example of deprecating a config file
option in a backwards compatible way.

Signed-off-by: Russell Bryant <rbryant@redhat.com>