e2e: Don't upgrade OS packages on AWS runners
The AMI we use is built to have some things pre-installed - most
notably the NVIDIA drivers and CUDA stack. I noticed that this upgrade
was now pulling in a new NVIDIA version, including a driver.

By not running an upgrade, we'll control when we apply updates to our
base image for important pre-installed dependencies.

Signed-off-by: Russell Bryant <rbryant@redhat.com>