fix: warn if vLLM is set and OS is not Linux
Running `ilab model serve --backend vllm` on macOS will never work since
vLLM is only supported on Linux.

The CLI now warns with:

```
$ ilab model serve --backend vllm
...
WARNING 2024-07-11 09:30:34,816 backends.py:195: get The serving backend 'vllm' was configured explicitly, but the provided model is not compatible with it. The model was detected as 'llama-cpp, reason: model is a GGUF file.'.
The backend startup sequence will continue with the configured backend
but might fail.
```

And the backend initialization will fail later on.

Signed-off-by: Sébastien Han <seb@redhat.com>