Debug: Improved display params diagnostics
The `display_params` decorator now adds two hidden diagnostic flags
`--debug-params` and `--debug-params-json` to the list of options.
When a flag is passed to a subcommand, it displays all parameters with
additional options and then exits with error code 0. The JSON output is
designed to be used in tests.

The function now displays more information, too. Besides name and value,
it reveals the type of the value and the origin (command line, env vars,
default, default, map).

```console
$ ilab model chat --debug-params
Parameters:
  question: ()         [type: tuple, src: commandline]
     model: 'models/merlinite-7b-lab-Q4_K_M.gguf'      [type: str, src: default_map]
   context: 'default'  [type: str, src: default_map]
   session: None       [type: None, src: default]
   ...
```

Signed-off-by: Christian Heimes <cheimes@redhat.com>