e2e: Add script option to skip training
I'm trying to bootstrap some test workflows using larger AWS instance
types. On these instances with multiple GPUs, it's not very useful to
test the legacy training code path, because it doesn't support using
multiple GPUs. On these instances, we want to test the new training
path using the training library. Integrating that is a work in
progress (PR #1557).

This is just a temporary measure that we can remove once the new
training code path is supported in this script.

Signed-off-by: Russell Bryant <rbryant@redhat.com>