e2e: Add a workflow that tests the full SDG pipeline
This workflow runs on a larger instance with 4 GPUs instead of 1. This
one is capable of serving Mixtral-8x7b-instruct (a quantized version,
at least). This is enough to test the full SDG pipeline.

The training library is targeted at systems with multiple GPUs, but we
don't have that in CI yet. For now, training is excluded from this
workflow. We will turn it on once that support is ready. See #1557 for
progress on that.

This commit also updates the CI docs to reflect what is tested by this
workflow.

Closes #1588

Signed-off-by: Russell Bryant <rbryant@redhat.com>