e2e: Always create a new config file
The previous code tried to only create a config file if it did not
already exist. This script has evolved to assume it has ownership of
the config, so we should always be creating it. We want to test that
part of the workflow, as well.

This came up when it was pointed out that the check for an existing
file was no longer valid after #1471.

Signed-off-by: Russell Bryant <rbryant@redhat.com>