feat: validate model_path from click
Instead of re-implementing the same logic of make sure the model_path
exists, is a file or a non-empty directory, we now have a Class that
does it for us.
This not only help with catching errors more quickly but also unifies
that approach accross various sub-commands like serve, train.

Co-authored-by: Christian Heimes <cheimes@redhat.com>
Signed-off-by: Sébastien Han <seb@redhat.com>