Fix security and UnboundLocalError in config show (PR #1471)
`ilab config show` mishandles an exception in YAML parsing. It does not
exit when `yaml.load()` fails with an exception. Instead it leaves
`config_yaml` undefined, which leads to an `UnboundLocalError` error.

The code was also using the wrong yaml loader. All other code paths are
using a safe loader, which prevents accidental code execution and
security problems. The FullLoader is not safe. PyYAML's documentation
even warns against using it:

> WARNING: As of pyyaml-5.3.1 there are still trivial exploits. Do
> not use this on untrusted data for now.

The command now exits with error code 2 on error and uses safe loading.
Our code is unlikely to be affected by the unsafe operation, but better
safe than sorry.

Signed-off-by: Christian Heimes <cheimes@redhat.com>