Allow for separate judge backend option
This is to allow for cases when for example your model or base model need to be served with llama-cpp but your judge needs to be served by vllm.  That would work as is if backend isn't specified.  But as soon as you need to force a backend, currently it would force the backend to be the same for the model and judge model

Signed-off-by: Dan McPherson <dmcphers@redhat.com>