Change --gpus to override when set and handle different vllmarg styles
Previously, --gpus wouldn't take effect if --tensor-parallel-size was set in serve vllm_args.  The normal behavior is that options should override config.  And a user could very possibly want to pass a different value than they use for the serve config.  Whether that value is set as an option or as evaluate config.

Also, the previous logic wasn't handling the --tensor-parallel-size=value format

Signed-off-by: Dan McPherson <dmcphers@redhat.com>