Add max workers recommendation for vllm
The recommendation is based on the number of cpus available since it relates to the number of cpu bound worker threads.  There is also some correlation to gpus as well and the results will vary based on hardware, so the recommended range is fairly wide +- 50% from the #cpus.

Also change the default max_workers in config to work better with lower end hardware.

Also switching usage of logging. to logger.

Signed-off-by: Dan McPherson <dmcphers@redhat.com>