Change vllm startup logic to not correlate retries to seconds
The list_models call takes over 2 seconds to respond which wasn't being accounted for in the timing of the desired sleep time.  The new logic changes the extra sleep to only 2 seconds, gets closer to a desired max wait time of 5 mins, and gives an accurate total duration on failure.

Signed-off-by: Dan McPherson <dmcphers@redhat.com>