ci: Run e2e-nvidia-t4-x1.yml automatically
This makes our smallest AWS-based e2e workflow run by default on PRs.
The hardware configuration is roughly equivalent to the existing e2e
job that runs by default on PRs on a github-hosted runner.

I propose running this in parallel with the existing e2e workflow for
a little while to evaluate its stability. The mergify configuraiton
has not been updated so it will not block on the results of this new
e2e variant.

One detail of this workflow that is very important to note is that it
is triggered on the `pull_request_target` event type instead of
`pull_request`. The difference is that the workflow runs using the
version checked in to `main` and does not use any code from the PR by
default. This is for security reasons. We do not want any secrets
exposed to code from PRs.

The middle job of this workflow, e2e, runs on an ephemeral AWS worker.
This workflow has no access to any secrets. This is the only place we
switch over to code in the PR. This should be safe, as long as we are
careful not to add any use of secrets to this part of the workflow.
All secret usage must be kept outside of the part that runs on the AWS
VM.

Signed-off-by: Russell Bryant <rbryant@redhat.com>