fix: ensure multiprocessing is used for distributed serving in vLLM
This commit introduces a dedicated function, `build_vllm_cmd` to
construct the vLLM command required to run the server. It ensures that
the command includes necessary arguments such as host, port, and model
path.

It also explicitly sets the distributed executor backend to
multiprocessing to avoid unintended use of Ray, even if it's installed
on the system. It uses the "--distributed-executor-backend mp" argument
to the vLLM command if it's not already present. With this, we ensure
that multiprocessing is used for distributed serving, even if Ray is
installed on the system, as Ray is not supported yet.

Signed-off-by: Sébastien Han <seb@redhat.com>