revert: use GGUF file as default model
There is a mismatch between the default model path set during `ilab
config init` and the one that the Pydantic (BaseModel) object holds.
Currently, the default model path is set to
`merlinite-7b-lab-Q4_K_M.gguf` in the `config.yaml` when we run `ilab
config init`. However, the config object holds the model path
`merlinite-7b-lab-Q4_K_M`. This discrepancy can lead to confusion and
potential errors when the model path is used in other parts of the
codebase. To resolve this issue, we should ensure that the default model
path set during `ilab config init` matches the one stored in the config
object. This will help maintain consistency and prevent any unexpected
behavior due to the mismatched model paths.

It's unclear why the default model was changed in
https://github.com/instructlab/instructlab/pull/1471. The commit message does
not provide any explanation for this change.
What was found is that originally `merlinite-7b-lab-Q4_K_M` was pointing
`DEFAULT_MODEL_OLD` and the `DEFAULT_MODEL` was
`models/merlinite-7b-lab-Q4_K_M.gguf`. `DEFAULT_MODEL_OLD` was only used
once in a chat function.

This adjustment is causing issues across various components, such as:

* `ilab model download` downloads the GGUF model.  * `ilab model serve` points
to the model directory.

As a result, unless we specify the model with `--model-path`, the serve command
does not work.

Now, the default model points to "merlinite-7b-lab-Q4_K_M.gguf" in the
config AND when running `ilab config init`.

Closes: https://github.com/instructlab/instructlab/issues/1736
Signed-off-by: Sébastien Han <seb@redhat.com>