utils: Drop reading of taxonomy data
Nothing uses the parsed taxonomy data that was returned by this code.
There is a proposal to change the taxonomy schema. Instead of fixing
this copy of this code, make it stop reading the contents aside from
doing automated validation.

Rename these functions to `validate_taxonomy` instead of
`read_taxonomy` since we only care about the validation part and not
returning any data from them.

If we look at changing this to use a new API, this helps clarify what
functionality we actually care about (validation, not the data).

Related to https://github.com/instructlab/sdg/issues/160

Signed-off-by: Russell Bryant <rbryant@redhat.com>