test: allow func test to use a specific temporary directory
Since #1471, the directory structure has changed and the functional test
was altered to setup the new directory and create a temporary directory
for the test data.  While this makes sense when running on a fresh new
runner, this makes the local test quite annoying since:

* a new temp is created
* a model is downloaded in this directory

Now, for local users, the script can be run like this:

```
TEST_DIR=/tmp/foo ./scripts/functional-tests.sh
```

This assumes you control your test directory and its cleanup.

Closes: https://github.com/instructlab/instructlab/issues/1774
Signed-off-by: Sébastien Han <seb@redhat.com>