SDG as input to Training e2e
feed the messages_* jsonl into training as the data source

bump the training lib to get the most recent deps.

Also, edit the a10x4 yml to set up the runner similarly to the a10x1.

flash-attn installation is broken as it requires packaging, wheel, and torch to pre-exist the instructlab-training[cuda] installation.
So, these three deps need to be added as an install before instructlab-training[cuda]

Signed-off-by: Charlie Doern <cdoern@redhat.com>