ci: Run the a10g-x4 workflow twice per day
This workflow tests the features that are the most difficult to test
in a personal development environment because they require the most
resources.

We're not ready to run it automatically on every commit for both
stability and cost reasons. Run it twice per day so we have a regular
stream of results based on what's in main.

It is also still possible to run it on demand on any PR or branch.

Signed-off-by: Russell Bryant <rbryant@redhat.com>