e2e: Drop all permissions for the untrusted portion of the workflow
In this workflow, we will be running untrusted code automatically on
PRs. That requires careful handling. For the portion that switches to
the PR branch and runs code, the `e2e` job that runs on the AWS VM,
ensure that the GITHUB_TOKEN has no permissions. We should also avoid
any use of any other secrets in this portion of the workflow. Right
now it is well isolated, but care must be taken to ensure it stays
that way.

Signed-off-by: Russell Bryant <rbryant@redhat.com>