evaluate: Demand --model argument through click library
The code makes assumptions that model is a string (meaning, it was
actually passed as --model argument). For example, it's passed into
`launch_server` that demands a string (not a None), and that will raise
a TypeError if None is passed.

The model being not None was (incompletely) validated inside
`get_evaluator`, but not consistently. This patch just adds
required=True to the click option, which allows to simplify the
validation code, and avoid some potential typing issues.

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>