Add support for an `ilab` global config
When the env var `ILAB_GLOBAL_CONFIG` is set, that config is used as the
template to render the ~/.config/instructlab/config.yaml user space
config. The interactive mode is skipped when calling `ilab config init`.
This works at the `ilab` main entrypoint, this can be used to specify a
different configuration file too when running any commands.

This ensures defaults like taxonomy repos, base models, etc, are honored
from the global config.

Co-authored-by: Sébastien Han <seb@redhat.com>
Signed-off-by: Charlie Doern <cdoern@redhat.com>