feat: update the dataset compatibility utility to support pretraining messages
In the future iterations of the SDG library, we plan to introduce a new dataset - phase007, which contains
a single message sample of the format "pretraining". This dataset is meant to pretrain models in the
new training library so that they can better retain knowledge documents. This commit allows the
new format to be compatible with the legacy dataset format by providing a conversion utility
which splits up the sample using the <|user|> and <|assistant|> tokens which are used when
delimiting the pretraining sample.

The conversion happens as follows: everything between the <|user|> and <|assistant|>
tokens is extracted and set to the "user" field of the legacy sample. Then, everything
following the "<|assistant|>"  token is interpreted as being the assistant token.

If the sample also included a system message as part of the messages, then this
is what gets used for the value of "system". Otherwise, an empty string is used for system.

This commit additionally moves a few tests which weren't covering the intended
portions correctly from the `test_lab_train.py` file and replaces them with
a more comprehensive series of tests in `test_utils.py`.

Signed-off-by: Oleg S <97077423+RobotSail@users.noreply.github.com>