introduce backends.common to reduce circular dependencies
The backends module, along with the llama_cpp and vllm modules,
have a circular dependencies, which is considered an anti-pattern."

This PR is intentionally kept small to facilitate an easier code review.
The plan is to relocate the relevant code from `backends` to `common`
in a subsequent PR, eliminating the circular dependency entirely.

References:
1. https://en.wikipedia.org/wiki/Circular_dependency
2. https://en.wikipedia.org/wiki/Acyclic_dependencies_principle

Signed-off-by: Costa Shulyupin <costa.shul@redhat.com>