tests: Don't access config file from $HOME/.local
The tests were running the CLI without --config option, which makes
`ilab` read from the default config location. Since we changes the
location of config and other files from "current dir" to ~/.local, it
becomes important to explicitly point the tests to (non-existing)
config to properly validate the tool behavior with a non-existing
config.

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>