fix: have chat andgenerate use the model being actively served
Presently, the `ilab model chat` and `ilab data generate` models require
the exact path to the model being actively served. This is a bad UX
since the path had to already have been specified correctly once
when it was initially being provided to `ilab model serve`.
Furthermore, this makes it very easy to get wrong and introduce new
bugs since we are having to propogate the model actively being used in serve.
This commit fixes that by automatically detecting the model ID and using that
for the model. This should also be fine because it is not expected that
the user will be serving more than one model at a time anyway.

Signed-off-by: Oleg S <97077423+RobotSail@users.noreply.github.com>