fix: allow for lora_qunatize_dtype to be null
There is a bug currently where lora_quantize_dtype under the training config cannot take on
values of null. The valid range of values for this field are either "nf4" or null,
however; the CLI only permits nf4 as valid input - ensuring that all models
will be quantize during training. This commit fixes that bug so now the valid
values will be "nf4" or null

Signed-off-by: Oleg S <97077423+RobotSail@users.noreply.github.com>