LoRA adjustments
add the ability to turn LoRA off. Currently all of the LoRA config fields are mandatory and have strict defaults meaning LoRA is always on when called via the CLI.
Also make it so that if LoRA is used with --is-padding-free=True, we fail. In testing these two do not work together

resolves #1906

Signed-off-by: Charlie Doern <cdoern@redhat.com>