fix: allow chat to resolve the local model being served
When `ilab model chat` is executed, it tries to find the default model being served
and use that when available. However; this behavior was not properly updated when
extending support for serving models via vLLM. As a result, the current code assumes that
the default model being served is the one which is specified as a default in the config.
Since users may instead download and serve a different model path, ilab model chat will
fail to detect and select this model when running ilab model chat.

This commit updates this behavior by introducing a default endpoint that ilab model serve
will commonly serve on (http://127.0.0.1:8000/v1), and checks to see if there is a
model being served at that endpoint when no other endpoint is specified. If the check
fails, then ilab model chat will proceed with the current behavior of seeking the default
model specified in the config for serve. If the check succeeds, then ilab model chat
simply proceeds using this endpoint for the remainder of the code.

Signed-off-by: Oleg S <97077423+RobotSail@users.noreply.github.com>