graduate --tensor-parallel-size
the vllm_args format: `-- --<flag_name>` hides commonly changed serve options behind an extra step. The commonly changed options also are not represented in our help text.
Graduate --tensor-parallel-size out of the hidden flags, rename it to `--gpus` and use the value of the flag unless `--tensor-parallel-size` is already set in the cfg.

resolves #1904

Signed-off-by: Charlie Doern <cdoern@redhat.com>