mergify config: Fix and update for valid regex and consistency
This fix updates the `files` attributes to use valid regex expressions
(`~=` operator) when needed and exact match (`=` operator) otherwise.
Globs are not supported and are replaced by regex.
We use `^` and `$` to properly anchor the regex as needed. Leading
and trailing `.*` are removed as they are not necessary.

For consistency across the multiple workflow conditions, we move the
`check-success` condition to the top. We also remove unneeded quote
characters as they were inconsistently used.

The `files` conditions are updated to properly match the globs used in
the corresponding workflows. References to no-longer-present files are
removed.

A `check-success` condition for the `spellcheck` job is added as it was
missing. An old action for release-v0.14 is removed.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>