feat: add more robust config defaulting
Currently,  one cannot specify a partial config. This is a useful thing to do when you don't care about some/most fields in the cfg, just GPU related ones.

pydantic has a `@model_validator` which allows a pre-check to be run before erroring out for missing fields. What I did in this commit is for each class and nested sub-class in `ilab`
the user can now pass little to no entires and we will fill out the rest of the config for them using the defaults in `get_default_config`. This differs from the current behavior which only allows the `Optional` fields to be omitted.

Signed-off-by: Charlie Doern <cdoern@redhat.com>