fix: get_model_family() fixes
This commit changes the following with
regard to get_model_family in configuration.py:

1. rename "forced" to "family" in the function
 to make it more readable.

2. get_model_family() was always guessing the
model_family even if a valid family was input.
Now if a valid family is input, it returns that.

3. add debug logging in backend.py to see the
result of get_model_family()

4. add a case in test_get_model_family() for when
model_family == "" which is it's default value
in the serve config and in other places in the
code base.

Signed-off-by: Ali Maredia <amaredia@redhat.com>