Increase default vllm startup attempts and make configurable
With larger models, the existing ~5 mins timeout proved not to be sufficient on some hardware.  The longest startup times seen so far have been in the 8 min range.  This moves the default up to 20 mins and makes the retry attempts configurable.  The new param is being added as optional so config changes aren't required.

Signed-off-by: Dan McPherson <dmcphers@redhat.com>