Fail fast when vllm errors out and exits
Previously, if vllm failed to start and raised an exception, the logic would have tried until its max attempts ran out to connect to a server that had already died.  This new logic detects if the server process is stopped, and if so outputs a message to use --enable-serving-output to debug further if the option is available.  So users will get a quick failure and guidance on debugging.

Signed-off-by: Dan McPherson <dmcphers@redhat.com>