fix: introduce vram stability waiting
vLLM requires GPU vram usage to be stable when calculating its memory
pools. A hard failure on start occurs when it detects this is not the
case. This creates a racy failure condition for various InstructLab
facilities, such as evaluation, which rely on multiple successive
restarts.

Address the problem by introducing a cool down period after process
termination. This will complete as early as 10s if VRAM stability
is detected. In the event of in-flight reclamation this will wait
as long as 30s.

Additionally address an issue where shutdown was being run twice.

Signed-off-by: Jason T. Greene <jason.greene@redhat.com>