fix: replace default endpoint with config value reading
Currently, when launching ilab model chat, we check to see if the user is serving
on the endpoint url before proceeding. The issue is that the endpoint url is set
to a default value equivalent to http://localhost:8000 - which also happens to
be the default value that we set in the serve config. Instead of checking for
this default, we should instead read the host port from the users serve config
and check that to see if anything is being served. This will provide cleaner
logic - as we can assume that the user just wants us to connect directly
when endpoint url is provided and skip the precheck which is currently used
to determine whether or not we should start a server ourselves

Signed-off-by: Oleg S <97077423+RobotSail@users.noreply.github.com>