fix: use the active server as the default endpoint url when generating
Today in ilab model serve, it is required to provide an endpoint url in order to override the
default behavior of ilab launching a separate model server to generate data. The other
alternative is to specify the model we are generating from, which is redundant since
the user had to have already done this when they initially served the model.

Therefore; this PR uses the users host port to check if a server at that address
is running, OpenAI-compliant, and actively serving a model. If all of these checks
succeed, we default to the user-provided endpoint URL when no other endpoint URL
is passed.

Signed-off-by: Oleg S <97077423+RobotSail@users.noreply.github.com>