hotfix: fix circular dependency issue
Currently we have an issue with a circular dependency inside the configuration code,
where we cannot include util code that imports from ilabclient in the utils module.
The reason why this happens is ilabclient also imports from configuration. But if we
have code which is technically a util, we cannot move it there for this reason.

Signed-off-by: Oleg S <97077423+RobotSail@users.noreply.github.com>