refactor: don't catch click.Error and then ctx.exit()
Bubbling the exception up has exactly the same effect - the return code
is the return code of the program.

See: https://click.palletsprojects.com/en/8.1.x/exceptions/#where-are-errors-handled

```
If a ClickException is raised, invoke the ClickException.show() method
on it to display it and then exit the program with
ClickException.exit_code.
```

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>