Refine wording of confirmation message for train profile re-init
Before the patch, the tool asked if the user wants to "continue" (which
may have been read as - continue the `init` process), and when the user
typed N, it would still recreate the config file.

The message is now more explicit, saying that the confirmation will only
apply to training profiles. The core config initialization will still
happen.

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>