hotfix: fix issue where default lora rank is overwriting setting of 0
Today in the map_train_to_library function, we parse and set the lora options conditionally.
However; the LoraOptions class from instructlab/training has a default rank of 4.
Since we are not explicitly setting this value, when the ilab user sets this to 0, it will
retain the default of 0. This commit fixes this issue by explicitly initializing the rank to 0.

Signed-off-by: Oleg S <97077423+RobotSail@users.noreply.github.com>