fix: reorder placement of profile reset prompt
In the current version of ilab, we check with the user when they reinitialize ilab that they are okay with
resetting their profiles directory. This check is the very first thing users see when they reinitialize their ilab setup.
The placement of this check may lead to confusion among users who may not be familiar with profiles if they are
new to instructlab. This PR moves the placement of the check to take place before the profile selection, so that
it is clearer what it may be doing.

Signed-off-by: Oleg S <97077423+RobotSail@users.noreply.github.com>