chore: remove unecessary @model_validator
All the defaults are coming from the Field(default), thus using the
@model_validator to populate default values is not needed anymore.

Also, the @model_validator is typically used for custom logic, like if
you have more complex logic for setting defaults, such as computing them
based on other fields or some external data, the model_validator could
be necessary. However, in this case, the logic is simply setting the
default values, which Pydantic can already do through Field().

A few options with default_factory were incorrect too, we don't need a
lambda here. The default_factory should directly reference the callable
that returns an instance of the class.

Co-Authored-By: Ihar Hrachyshka <ihrachys@redhat.com>
Signed-off-by: Sébastien Han <seb@redhat.com>