scripts/infra: Print help when only cloud type is passed
When running the following command,

    cloud-instance.sh ec2

the output would be empty as the script would silently exit on the
`shift`. It is desirable to output the usage help text in this case.
Adjust the `ec2` and `ibm` handlers to handle this case and print the
usage text when no command is provided.

Signed-off-by: Russell Bryant <rbryant@redhat.com>