Add better tuning for max_workers based on perf tests
The result of those tests showed ideal values for max_workers needs to take gpu and cpu counts into account.  The new recommendation does have a magic number but hopefully the suggestion range is large enough to account for the magic number potentially not being perfect.  Removing the magic number would require understanding the relative compute power of the gpus.

Signed-off-by: Dan McPherson <dmcphers@redhat.com>