Update A100_H100_x4 to slightly lower max_batch_len (55000->54000)
Essentially with training, we have found that the maximum value we can push that max_batch_len field to without the chance of hitting an OOM lies somewhere between 54k and 55k. A batch with a size above that value will overflow an 80GB card (in a 4x scenario) and crash with an OOM. Anything below that value should be within 80GB. Originally we thought that cut-off was 55k, but we realized it was slightly lower, so adjusting accordingly.

Signed-off-by: Mustafa Eyceoz <meyceoz@redhat.com>