taxonomy diff/data generate: support taxonomy-base=empty
This allows users to specify that they want all taxonomy files in
their repo to be included, where 'empty' symbolically represents
the empty tree before this repo was created.

The latest version of instructlab-sdg adds support for this behavior,
but we have a copy of this code here for taxonomy diff also.

Co-authored-by:Dan McPherson <dmcphers@redhat.com>
Signed-off-by: Mark McLoughlin <markmc@redhat.com>