Retry vllm startup for long running processes
The reason for adding retries is instructlab has to start and stop vllm a lot of times to work through its workflow end to end.  Whenever a component is susceptible to intermitent failures (which vllm seems to be with some hardware configurations) it risks the long running process failing as a whole.  Or to the last savepoint as much as they exist.  The goal of this logic is to retry vllm startups for all the startups that exist in long running processes.  For now the logic is set to retry once for generate, phased training, and evaluate.

Once we have some level of confidence this approach is reasonable and we do want a value in config, there is some concern that config already has a variable named max_startup_attempts which should really be something like max_startup_connect_attempts or max_avail_checks since it is checking whether vllm is running.  This new variable might be called max_startup_retries.  And until we went through the process to change the current config var, we would just have to know what the two variables map to.

We could have an env var to set max_startup_retries, but then technically we would need one for generate and one for evaluate.  Or one value would apply globally to all vllm starts.  Or one value would apply globally but only to evaluate, phased training, and generate.

Point being it's a bit messy and the provided impl of defaulting to 1 retry for generate and evaluate seems reasonable for all.

Signed-off-by: Dan McPherson <dmcphers@redhat.com>
(cherry picked from commit da135b63f8e071493a4cdf7a40ba9eeb20ed214c)